Bandera_tricolor

Description

El problema de la bandera tricolor consiste en lo siguiente: Dada un lista de objetos xs que pueden ser rojos, amarillos o morados, se pide devolver una lista ys que contiene los elementos de xs, primero los rojos, luego los amarillos y por último los morados.

Definir el tipo de dato Color para representar los colores con los constructores R, A y M correspondientes al rojo, azul y morado y la función

banderaTricolor :: [Color] -> [Color]

tal que (banderaTricolor xs) es la bandera tricolor formada con los elementos de xs. Por ejemplo,

>>> banderaTricolor [M,R,A,A,R,R,A,M,M]
[R,R,R,A,A,A,M,M,M]
>>> banderaTricolor [M,R,A,R,R,A]
[R,R,R,A,A,M]

Documentation

data Color

banderaTricolor :: [Color] -> [Color]

verifica_banderaTricolor :: IO ()

Comprueba la equivalencia de las definiciones

>>> verifica_banderaTricolor
+++ OK, passed 100 tests.